Fix TCP transport with IPv6 addresses
Replace AF_INET with socket.getaddrinfo() to resolve the correct address
family for the given address.

Only call ip_bracket() when the brackets are actually required, instead
of always by default. Raw IPv6 addresses are required most of the time,
and brackets are only required when representing the address with the
port, or as a part of a URI.

Use the provided address family inside TCPClientKeepAlive
_create_stream() instead of blindly assuming AF_INET

Signed-off-by: Joe Groocock <me@frebib.net>