HACK: always sync pillar extmods before loading
This is required when the saltenv points to a gitfs backend and the
extmods haven't been sync'ed before running. It would be good if Salt
could do this automatically when running against a new saltenv, and if
it kept the extmods cache separate for each saltenv to prevent
potentially racy behaviour.

Signed-off-by: Joe Groocock <me@frebib.net>