Implement compile_template in pillar stack
Replace the existing hard-coded Jinja environment and YAML parsing with
Salt built-in templates.compile_template functionality. It allows for
multiple types of templates, not just jinja|yaml. It also keeps the
Jinja environment/configuration aligned with that used in normal sls
templates.

This change has some functional differences, of which the most important
and breaking are:
- Jinja templates are rendered as normal sls would be, with
  StrictUndefined being the default, unless that option is otherwise
  overriden in the minion/master opts.
- Jinja global dunders __salt__, __opts__, __grains__, __pillar__ etc
  are now referenced as normal variables: salt, opts, grains, pillar
  with the exception of __stack__ and stack being independent entities.
  __stack__ remains as a dict of template functions to remain consistent
  with the existing pillar stack behaviour.

Signed-off-by: Joe Groocock <jgroocock@cloudflare.com>