SALT-237: Allow configurable pillar rendering timeout
The ZMQ transport defaults to making 1+3 attempts with a 60s timeout each, which is actually misleading as the ReqChannel does not actually send another request before a response is received - so it's better understood as a single 4x60s=240s attempt.
In any case, the remote pillar rendering function currently does not override or make this configurable, which causes any pillar render above 4m to always fail without the ability to tune this behavior. We have nodes whose pillar render is consistently above this threshold (trimming for another day), and building in the option seems sensible.