states/file: fix mis-reported removed files with clean
file.directory with clean will remove any unamanged files that it
encounters. Directories that contain files nested in child directories
were not accounted for, even though they are in fact not removed. The
discrepancy originates from the two-part code where files that are "to
be removed" are reported initially, then re-enumerated when actually
being removed. The on-disk behaviour is correct; it is only the changes
dict which is misreported.

Below is a minimal test reproducing the issue where the file is claimed
to be removed, but is actually not removed at all:

    local:
    ----------
              ID: /tmp/somedir/otherdir/file
        Function: file.managed
          Result: True
         Comment: File /tmp/somedir/otherdir/file is in the correct state
         Started: 15:08:42.705800
        Duration: 26.253 ms
         Changes:
    ----------
              ID: /tmp/somedir
        Function: file.directory
          Result: True
         Comment: Directory /tmp/somedir updated
         Started: 15:08:42.732313
        Duration: 2.705 ms
         Changes:
                  ----------
                  /tmp/somedir/otherdir:
                      ----------
                      removed:
                          Removed due to clean

The source SLS being:

    /tmp/somedir:
      file.directory:
        - clean: true
        - makedirs: true

    /tmp/somedir/otherdir/file:
      file.managed:
        - contents: some words
        - makedirs: true
        - require_in:
          - file: /tmp/somedir

Signed-off-by: Joe Groocock <jgroocock@cloudflare.com>