fileclient: Save one round-trip by skipping unused stat
get_file() requests both the hash and mode(stat) of a file from the
master before verifying the local cache. The stat appears to be unused
resulting in a wasted round-trip to the master, which can be costly in
high-latency environments.

The stat request was added in 73a156d28cafd45dcc6c2aac9f1ac3ad905695e3
and the usage of the stat values removed again shortly after in
6e34c2b5e5e849302af7ccd00509929c3809c658 but the stat call remained.

Signed-off-by: Joe Groocock <jgroocock@cloudflare.com>