Implement get_files() for fetching many files in batch
Enumerate and fetch many files from the salt-master in a single request,
only receiving what is needed compared to the local filesystem cache.
The fileclient sends a hash for every file it already has, and the
master will only return files that have changed to avoid retransmitting
data needlessly.

This is VERY WIP. DO NOT USE!?!L?!?!?!

Signed-off-by: Joe Groocock <jgroocock@cloudflare.com>