Saltstack versions can have cf? suffix
Allows identifying internal Cloudflare Salt builds versus upstream
releases. Set environment SALT_CF_VERSION to get a Cloudflare version
and SALT_CF_RELEASE to get a release version without the git sha.

Base non-release cf versions on the previous commit. A release build
version should be greater than the non-production commit that created
it. Currently we use the same cf revision in the dev version string as
in the release version string, which breaks that contract. To mitigate
this, non-release versions will be based on the cf-revision of the
*previous* commit count; for example:

    v300x-cf10.g1234abcd0 when released would produce v300x-cf11

as it is 11 commits ahead of the v300x tag.

This obviously doesn't take into account a dev build that is multiple
commits ahead of the last "actual release", because we don't track which
commits make it as a release or not. In those cases the dev package
should still have a version number that is strictly "less than" the
release version string though.

Signed-off-by: Joe Groocock <jgroocock@cloudflare.com>