Mirror _check_dir_meta in check_perms by allowing user/group in name and ID format
The issue here is that the salt state `file.directory` eventually uses
`_check_dir_meta` to determine pending state changes for ownership and
permissions. This function accepts user/group in both ID and name form,
and will return the corresponding changes. However salt uses
`file.check_perms` to actually make the changes, and in this function
salt converts IDs into names, and then determines whether any changes
are necessary. We're currently in the situation where users can be
mapped to two different IDs depending on which nsswtich source is used.
This means we need to compare IDs instead of names when setting file
ownership, so the behaviour of `_check_file_meta` is preferred.

This also ends up fixing a couple issues in the post ownership change
verification where conditionals `user != ""` are used in the group
verification section.

A new test for checking that both user and group can be supplied by name
or ID to `file.check_perms` and it will correctly identify if any do not
match the current file permissions.