SALT-552 Build salt-omni package with Tiamat
Using the same source tree, build a onedir-style package ala v3005
upstream packaging but with a from-source build of python 3.9 to work on
older systems. This is intended as an upgrade path for legacy systems
running EOL builds of Debian such as Stretch and Buster.

This replaces salt-common as well as salt-minion/salt-master and other
similar packages by providing them all in a single binary/directory
bundled with the python interpreter and everything it needs to work.

EOL distros like Stretch/Buster provide python3 that are too old and
don't work at all/well with Salt >=3005 so instead we build a known-good
Python version for consistency, which tiamat picks up and bundles into
the onedir package.

Signed-off-by: Joe Groocock <jgroocock@cloudflare.com>